add "Unmasked areas" option
